Added Start and End Navigation functions
MapX390-44
Mapx390-39

Added two new functions that are callable by Android, when a navigation
is to start/end.

startNavigation is called either when a storyline is commenced or after
a call to Android is sent from navigateToPOI()

endNavigation can be called when a storyline or a single navigation
(POI to POI) is cancelled

Currently, when either function is called, the only thing is the change
of the start and end POI icons. The icons will be changed to another
color to emphasize the POIs.

Also did a small refactoring to add the navigateToPOI function within
the Controller class.