Change to positioning logic
MAPX390-45

Change to positioning logic within the map.

The old map was basically a cartesian plan with only the first
quadrant. The new logic is that all four quadrants are used, with the
map/floor being centered. This is the reason why there are now things
like offsets, since MapEditor will send us coords where (0,0) will be
from a corner.

With this new map logic, I can extend the map frame all four sides so
that the popup cutoff does not happen anymore.